fix(workbench): make dock previews reliable
