+++ apps/desktop/src/main/host/tuttiAssetProtocol.ts
@@ -2,6 +2,7 @@ import { pathToFileURL } from "node:url";
 import { app, net, session, type Session } from "electron";
 import { tuttiAssetProtocolScheme } from "../../shared/tuttiAssetProtocol.ts";
 import { resolveTuttiAssetProtocolFilePath } from "./tuttiAssetProtocolResolver.ts";
+import { createTuttiAssetProtocolResponse } from "./tuttiAssetProtocolResponse.ts";
 
 const handledSessions = new WeakSet<Session>();
 
@@ -24,6 +25,8 @@ export function registerTuttiAssetProtocolForSession(
     if (!filePath) {
       return new Response(null, { status: 404 });
     }
-    return net.fetch(pathToFileURL(filePath).href);
+    return createTuttiAssetProtocolResponse(
+      await net.fetch(pathToFileURL(filePath).href)
+    );
   });
 }
